+++ fedmdrs_main.py
@@ -41,7 +41,7 @@ def fedmdrs_main(
         train_clients = get_PSM_train_clients(num_clients, required_length=trans_len)
         # reduce train data size to investigate the effect of data size
         train_clients = [
-            train_data[: data_proportion * len(train_data)]
+            train_data[: int(data_proportion * len(train_data))]
             for train_data in train_clients
         ]
