+++ tranad_main.py
@@ -51,7 +51,7 @@ def tranad_main(
     else:
         train_data = get_PSM_train()
         # reduce train data size to investigate the effect of data size
-        train_data = train_data[: data_proportion * len(train_data)]
+        train_data = train_data[: int(data_proportion * len(train_data))]
 
         test_clients = get_PSM_test_clients()
     n_features = train_data.shape[1]
